fix: Split implementation to smaller modules.
