+++ tools/forms/libs/form-ui/core/form-generator/path-utils.js
@@ -1,0 +1,17 @@
+/**
+ * Path/ID helpers used across Form UI
+ */
+
+export function hyphenatePath(path) {
+  return String(path || '').replace(/[.\[\]]/g, '-');
+}
+
+export function pathToGroupId(path) {
+  return `form-group-${hyphenatePath(path)}`;
+}
+
+export function arrayItemId(arrayPath, index) {
+  return `form-array-item-${hyphenatePath(arrayPath)}-${index}`;
+}
+
+
