+++ tools/forms/libs/form-ui/core/inputs/select-input.js
@@ -1,0 +1,24 @@
+import BaseInput from './base-input.js';
+
+export default class SelectInput extends BaseInput {
+  create(fieldPath, enumValues, propSchema) {
+    const select = document.createElement('select');
+    select.name = fieldPath;
+    select.className = 'form-ui-select';
+    const emptyOption = document.createElement('option');
+    emptyOption.value = '';
+    emptyOption.textContent = '-- Select --';
+    select.appendChild(emptyOption);
+    enumValues.forEach((value) => {
+      const option = document.createElement('option');
+      option.value = value;
+      option.textContent = value;
+      if (propSchema.default === value) option.selected = true;
+      select.appendChild(option);
+    });
+    this.attachCommonEvents(select, fieldPath, propSchema);
+    return select;
+  }
+}
+
+
